requirements.txt
# Requirements for xNetVN Monitor Daemon

# Core dependencies
PyYAML>=6.0
psutil>=5.9.0

# Optional dependencies for advanced features
# requests>=2.31.0  # For HTTP-based health checks
# prometheus-client>=0.19.0  # For Prometheus metrics export
